renamed SfdxOrgConfig.py to sfdx_org_config.py; removed sfdx property This was done to make it easier to create mocks for testing
